chore(team): split remote relay tests
